switch to meson and organise tree
The addition of libftdi support breaks distro portability as some
distributions use the name "libftdi" and other use "libfdti1". It felt
like a good enough reason to spend a little time upgrading the build
system, as while this isn't that hard to solve with make, the growing
number of files and general complexity of the project invited some
cleanup.